serverless-layers/lambda-layer-odbc/docker-build/requirements.txt
boto3==1.17.53
botocore==1.20.53
certifi==2020.12.5
chardet==4.0.0
idna==2.10
jmespath==0.10.0
python-dateutil==2.8.1
requests==2.25.1
s3transfer==0.3.7
six==1.15.0
urllib3==1.26.4
pyodbc==4.0.30
pymongo==3.11.4
pyhumps==1.6.1
